requirements.txt
colorama
pydbus
pillow
sounddevice
numpy
pydbus
